+++ js/utils/change-attribute-selected.js
@@ -1,0 +1,10 @@
+const setAttributeSelected = (item) => {
+  item.setAttribute('selected', 'selected');
+  return item;
+};
+
+const removeAttributeSelected = (item) => {
+  item.removeAttribute('selected', 'selected');
+};
+
+export {setAttributeSelected, removeAttributeSelected};
